Types

FEM machines

Nonlinear deformation

FinEtoolsDeforNonlinear.FEMMDeforNonlinearBaseModule.AbstractFEMMDeforNonlinear — Type

AbstractFEMMDeforNonlinear <: AbstractFEMMDeforLinear

Abstract type of FEMM for nonlinear deformation.

FinEtoolsDeforNonlinear.FEMMDeforNonlinearModule.FEMMDeforNonlinear — Type

FEMMDeforNonlinear{MR<:AbstractDeforModelRed,  S<:AbstractFESet, F<:Function, M<:AbstractMatDeforNonlinear} <: AbstractFEMMDeforNonlinear

Class for nonlinear deformation finite element modeling machine.

FinEtoolsDeforNonlinear.FEMMDeforNonlinearModule.FEMMDeforNonlinear — Method

FEMMDeforNonlinear(mr::Type{MR}, integdomain::IntegDomain{S, F}, material::M) where {MR<:AbstractDeforModelRed, S<:AbstractFESet, F<:Function, M<:AbstractMatDeforNonlinear}

Constructor of nonlinear deformation finite element modeling machine.

FinEtoolsDeforNonlinear.FEMMDeforNonlinearModule.FEMMDeforNonlinear — Method

FEMMDeforNonlinear(mr::Type{MR}, integdomain::IntegDomain{S, F}, material::M) where {MR<:AbstractDeforModelRed, S<:AbstractFESet, F<:Function, M<:AbstractMatDeforNonlinear}

Constructor of nonlinear deformation finite element modeling machine.

The material coordinate system is the default.

Material models

Material models for nonlinear deformation

FinEtoolsDeforNonlinear.MatDeforNonlinearModule.AbstractMatDeforNonlinear — Type

AbstractMatDeforNonlinear <: AbstractMatDefor

Abstract nonlinear material.

Material models for neohookean hyperelasticity

FinEtoolsDeforNonlinear.MatDeforNeohookeanModule.MatDeforNeohookean — Type

MatDeforNeohookean{MR<:AbstractDeforModelRed, MTAN<:Function, MUPD<:Function} <: AbstractMatDeforNonlinear

Type for triaxial neohookean hyperelastic material.

Note

The material object is not thread safe. It holds temporary arrays. If the object is to be used in a multi-threaded environment, each thread must have its own private copy.

FinEtoolsDeforNonlinear.MatDeforNeohookeanModule.MatDeforNeohookean — Method

MatDeforNeohookean(mr::Type{DeforModelRed3D}, mass_density::FFlt, E::FFlt, nu::FFlt)

Create triaxial neohookean hyperelastic material.

FinEtoolsDeforNonlinear.MatDeforNeohookeanModule.MatDeforNeohookean — Method

    MatDeforNeohookean(mr::Type{MR}, E::FFlt, nu::FFlt) where {MR}

Create neohookean isotropic elastic material.

The mass density is the default value.

Material models for St Venant-Kirchhoff hyperelasticity

FinEtoolsDeforNonlinear.MatDeforStVKModule.MatDeforStVK — Type

MatDeforStVK{MR<:AbstractDeforModelRed, MTAN<:Function, MUPD<:Function} <: AbstractMatDeforNonlinear

Type for triaxial St Venant-Kirchhoff hyperelastic material.

FinEtoolsDeforNonlinear.MatDeforStVKModule.MatDeforStVK — Method

MatDeforStVK(mr::Type{DeforModelRed3D}, mass_density::FFlt, E1::FFlt, E2::FFlt, E3::FFlt, G12::FFlt, G13::FFlt, G23::FFlt, nu12::FFlt, nu13::FFlt, nu23::FFlt)

Create triaxial St Venant-Kirchhoff hyperelastic material.

In general, the material is assumed to be orthotropic. There is a specialized constructor for an isotropic version.

FinEtoolsDeforNonlinear.MatDeforStVKModule.MatDeforStVK — Method

MatDeforStVK(mr::Type{MR}, E::FFlt, nu::FFlt) where {MR}

Create St Venant-Kirchhoff isotropic elastic material.

The mass density is the default value.

FinEtoolsDeforNonlinear.MatDeforStVKModule.MatDeforStVK — Method

MatDeforStVK(mr::Type{MR}, E::FFlt, nu::FFlt) where {MR}

Create St Venant-Kirchhoff isotropic elastic material.

The mass density is the default value.